ai-novel-editor/requirements-dev.txt
# ============================================================================
# AI Novel Editor - Development Requirements
# ============================================================================
# This file includes all dependencies for development, testing, and optional
# features.
# Install with: pip install -r requirements-dev.txt
# ============================================================================

# Include base requirements
-r requirements.txt

# Development & Testing
# ----------------------------------------------------------------------------
pytest>=7.0.0,<8.0.0            # Testing framework
pytest-asyncio>=0.21.0          # Async test support
pytest-cov>=4.0.0               # Coverage reporting
black>=23.0.0                   # Code formatter
isort>=5.12.0                   # Import sorter
flake8>=6.0.0                   # Linting
mypy>=1.0.0                     # Type checking
pylint>=2.17.0                  # Code analysis

# Google AI Integration (recommended for full functionality)
# ----------------------------------------------------------------------------
google-genai>=0.3.0             # Google Gemini API client

# Optional: Advanced Features
# ----------------------------------------------------------------------------
# Uncomment if you want to develop these features:
# streamlit-agraph>=0.0.45      # Graph visualization
# plotly>=5.0.0                 # Interactive charts
# networkx>=3.0                 # Graph data structures
